--- Microsoft.Toolkit.Graph/RoamingSettings/RoamingSettings.cs
@@ -1,11 +1,0 @@
-﻿namespace Microsoft.Toolkit.Graph.RoamingSettings
-{
-    public class RoamingSettings : UserExtensionBase
-    {
-        public override string ExtensionId => "com.contoso.roamingSettings";
-
-        public RoamingSettings(string userId) : base(userId)
-        {
-        }
-    }
-}
