+++ SampleTest/Samples/RoamingSettings/RoamingSettingsView.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="SampleTest.Samples.RoamingSettings"
+    x:Class="SampleTest.Samples.RoamingSettings.RoamingSettingsView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:SampleTest.Samples"
+    xmlns:local="using:SampleTest.Samples.RoamingSettings"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:graph="using:Microsoft.Graph"
