+++ SampleTest/Samples/RoamingSettings/RoamingSettingsView.xaml.cs
@@ -1,15 +1,19 @@
-﻿using Windows.UI.Xaml.Controls;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
-namespace SampleTest.Samples
+using Windows.UI.Xaml.Controls;
+
+namespace SampleTest.Samples.RoamingSettings
 {
     /// <summary>
     /// A sample for demonstrating features in the RoamingSettings namespace.
     /// </summary>
-    public sealed partial class RoamingSettings : Page
+    public sealed partial class RoamingSettingsView : Page
     {
         private RoamingSettingsViewModel _vm => DataContext as RoamingSettingsViewModel;
 
-        public RoamingSettings()
+        public RoamingSettingsView()
         {
             InitializeComponent();
             DataContext = new RoamingSettingsViewModel();
