+++ SampleTest/Samples/RoamingSettings/RoamingSettingsViewModel.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.Graph;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.Graph;
 using Microsoft.Toolkit.Graph.Providers;
 using Microsoft.Toolkit.Graph.RoamingSettings;
 using System.Collections.Generic;
@@ -7,7 +11,7 @@ using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 
-namespace SampleTest.Samples
+namespace SampleTest.Samples.RoamingSettings
 {
     public class RoamingSettingsViewModel : INotifyPropertyChanged
     {
